+++ taiga/users/migrations/0011_user_theme.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+from __future__ import unicode_literals
+
+from django.db import models, migrations
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('users', '0010_auto_20150414_0936'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='user',
+            name='theme',
+            field=models.CharField(null=True, blank=True, max_length=100, default='', verbose_name='default theme'),
+            preserve_default=True,
+        ),
+    ]